refactor: migrate to typescript and SQIP v1
