+++ packages/gatsby-transformer-sqip/.gitignore
@@ -1,4 +1,5 @@
-/*.js
-!index.js
+dist
+*.js
+!gatsby-node.js
 yarn.lock
 package-lock.json
