+++ packages/gatsby-transformer-sqip/gatsby-node.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require("./dist/gatsby-node");
