+++ packages/gatsby-transformer-sqip/index.js
@@ -1,1 +1,1 @@
-// noop
+module.exports = require(`./dist/index.js`)
