--- packages/gatsby-transformer-sqip/src/gatsby-node.js
@@ -1,1 +1,0 @@
-exports.setFieldsOnGraphQLNodeType = require(`./extend-node-type`)
