+++ packages/gatsby-transformer-sqip/src/gatsby-node.ts
@@ -1,0 +1,2 @@
+exports.setFieldsOnGraphQLNodeType =
+  require(`./set-fields-on-graph-ql-node-type`).setFieldsOnGraphQLNodeType
