+++ packages/gatsby-transformer-sqip/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": [
+    "node_modules",
+    "src/__tests__",
+    "src/__mocks__",
+    "dist",
+    "./node.d.ts"
+  ]
+}
